requirements.txt
requests==2.32.0
pytest==7.4.0
matplotlib
pandas
colorama
reportlab==4.0.4
jinja2==3.1.2
PyPDF2==3.0.1
beautifulsoup4==4.12.2
weasyprint==59.0
pyyaml==6.0.2